Fix signedness.
2001-05-13  Almer S. Tigelaar  <almer@gnome.org>

	* sc.c (sc_parse_coord): Fix signedness.

+++ plugins/sc/sc.c
@@ -99,7 +99,7 @@ static void
 sc_parse_coord (const char **strdata, int *col, int *row)
 {
 	const char *s = *strdata, *eq;
-	size_t len = strlen (s);
+	int len = strlen (s);
 	char tmpstr [16];
 
 	g_return_if_fail (strdata);
